requirements.txt
anthropic
python-dotenv
mcp>=0.1.0